Completed draft exploration with original ds7 10.000 point dataset
It took 8 hours to complete, and it didn't correctly identify the clusters. According to an original version of the paper, the number of subclusters to create would need to be between 1% to 5% (n=500) the number of points. However when running with n=500, it seems it will take 135 hours+, and it might also need to adjust knn and alpha to get a good result.
Two options: adjusting the algorithm so that it is more efficient (1 day of work maybe?), or work with smaller datasets (focus on the scikit learn ones). For the moment going for the second option.